LayoutTests/media/modern-media-controls/buttons-container/buttons-container-layout.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/buttons-container.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/buttons-container.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>ButtonsContainer</code> layout.");

const tenPtWideButton = new Button;
tenPtWideButton.width = 10;

const twentyPtWideButton = new Button;
twentyPtWideButton.width = 20;

const thirtyPtWideButton = new Button;
thirtyPtWideButton.width = 30;

// Should be disregarded by the container.
const disabledButton = new Button;
disabledButton.width = 15;
disabledButton.enabled = false;

// Should be disregarded by the container.
const droppedButton = new Button;
droppedButton.width = 25;
droppedButton.dropped = true;

const container = new ButtonsContainer({
    margin: 10,
    padding: 20,
    buttons: [tenPtWideButton, disabledButton, twentyPtWideButton, droppedButton, thirtyPtWideButton]
});

container.layout();
shouldBe("container.children.length", "3");
shouldBe("container.children[0]", "tenPtWideButton");
shouldBe("container.children[1]", "twentyPtWideButton");
shouldBe("container.children[2]", "thirtyPtWideButton");
shouldBe("tenPtWideButton.x", "20");
shouldBe("twentyPtWideButton.x", "40");
shouldBe("thirtyPtWideButton.x", "70");
shouldBe("container.width", "120");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
